etl-pipeline/requirements.txt
google-cloud-discoveryengine==0.11.0
google-cloud-logging==3.8.0
google-cloud-aiplatform==1.38.1
msal==1.26.0
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
pydantic==2.5.3 # For robust data models